Pulled out FieldGroup tests into own class, added error case coverage
